[publishing] 메인치매예방섹션, 서브치매예방섹션 추가, 컴포넌트 분리
